requirements.txt
django==3.0
djangorestframework==3.12
djangorestframework-jwt==1.11.0
drf-haystack==1.8.11
elasticsearch==7.6.0
channels==2.3
channels-redis==4.0.0
paho-mqtt==1.6.1
gunicorn==20.1.0
pymysql==1.0.2
supervisor==4.2.5
daphne==2.5.0